requirements.txt
# Core Trading and API
alpaca-py>=0.21.0
requests>=2.31.0
aiohttp>=3.8.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Configuration and environment management
pyyaml>=6.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Visualization and plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Data export and file handling
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Statistical analysis
statsmodels>=0.14.0

# Performance optimization
numba>=0.58.0

# Logging and monitoring
structlog>=23.0.0
colorlog>=6.7.0

# CLI and utilities
click>=8.1.0
tqdm>=4.66.0
tabulate>=0.9.0

# Testing framework
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Jupyter notebook support (optional for analysis)
jupyter>=1.0.0
ipython>=8.0.0

# Web interface (optional)
streamlit>=1.28.0

# Additional math and finance libraries
quantlib>=1.32.0;platform_system!="Windows"
yfinance>=0.2.18

# Memory profiling (development)
memory-profiler>=0.61.0

# API rate limiting
ratelimit>=2.2.1

# Time series analysis
ta-lib>=0.4.28;platform_system!="Windows"

# Notifications (optional)
plyer>=2.1.0